requirements.txt
langchain
langchain-community
unstructured[all-docs]
faiss-cpu
transformers
torch
sentence-transformers
accelerate
